+++ guillotina/tests/test_api.py
@@ -516,3 +516,11 @@ async def test_create_content_fields(container_requester):
         assert len(response['categories']) == 1
         assert response['textline_field'] == 'foobar'
         assert response['text_field'] == 'foobar'
+
+
+async def test_raise_http_exception_works(container_requester):
+    async with container_requester as requester:
+        response, status = await requester('POST', '/@raise-http-exception')
+        assert status == 422
+        response, status = await requester('GET', '/@raise-http-exception')
+        assert status == 422